+++ Engine/MathUtils.cpp
@@ -12,7 +12,7 @@ ImVec2 MathUtil::ScaleVec2(ImVec2 input) {
 }
 
 ImVec2 MathUtil::ScaleVec2(double x, double y) {
-	ImVec2 input(x, y);
+	ImVec2 input((float)x, (float)y);
 
 	Window* wnd = Window::GetInstance();
 	if (wnd->IsScaleOutput()) {
